fix: Fix type compatibility issues for macOS ARM64
- Change size_t to uint64_t in free_tile() signature for godot-cpp
  compatibility (size_t maps to unsigned long on macOS which isn't
  supported by PtrToArg)
- Fix designated initializer in CesiumPropertyInfo which fails on
  non-aggregate types with Apple Clang

+++ cesium_godot/Implementations/GodotPrepareRenderResources.cpp
@@ -144,9 +144,9 @@ void* GodotPrepareRenderResources::prepareInMainThread(Tile& tile, void* pLoadTh
 }
 
 void GodotPrepareRenderResources::free(Tile& tile, void* pLoadThreadResult, void* pMainThreadResult) noexcept
-{	
+{
 	const auto& tileId = tile.getTileID();
-	size_t hash = std::visit(CesiumVariantHash{}, tileId);
+	uint64_t hash = static_cast<uint64_t>(std::visit(CesiumVariantHash{}, tileId));
 	auto* instance = static_cast<Cesium3DTile*>(pMainThreadResult);
 	this->m_tileset->call_deferred("free_tile", instance, hash);
 }
